Refactor tests and dependencies - upgrade to junit 5 annotations - introduce jsr-330 annotations
