+++ build-dependencies.gradle
@@ -41,10 +41,12 @@ dependencies {
 
   annotationProcessor group: 'org.projectlombok', name: 'lombok', version: versions.lombok
 
-  compile (group: 'uk.gov.hmcts.reform.auth', name: 'auth-checker-lib', version: '2.1.4') {
+  implementation (group: 'uk.gov.hmcts.reform.auth', name: 'auth-checker-lib', version: '2.1.4') {
     exclude group: 'uk.gov.hmcts.reform', module: 'java-logging-spring'
   }
-  compile group: 'uk.gov.hmcts.reform', name: 'logging-appinsights', version: versions.logging
+  implementation group: 'uk.gov.hmcts.reform', name: 'logging-appinsights', version: versions.logging
+
+  implementation group: 'javax.inject', name: 'javax.inject', version: '1'
 
 
 // TODO this excludes the following jars, but the spring boot template does not
