+++ src/main/java/uk/gov/hmcts/reform/em/hrs/controller/RootUrlController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.em.hrs.controllers;
+package uk.gov.hmcts.reform.em.hrs.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
