+++ src/main/java/uk/gov/hmcts/reform/em/hrs/service/FolderService.java
@@ -1,20 +1,23 @@
 package uk.gov.hmcts.reform.em.hrs.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.reform.em.hrs.domain.Folder;
 import uk.gov.hmcts.reform.em.hrs.repository.FolderRepository;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.util.Optional;
 import java.util.UUID;
 
 @Transactional
-@Service
+@Named
 public class FolderService {
+    private final FolderRepository folderRepository;
 
-    @Autowired
-    private FolderRepository folderRepository;
+    @Inject
+    public FolderService(FolderRepository folderRepository) {
+        this.folderRepository = folderRepository;
+    }
 
     public Optional<Folder> findById(UUID id) {
         return folderRepository.findById(id);
