+++ src/test/java/uk/gov/hmcts/reform/em/hrs/controller/RootUrlControllerTest.java
@@ -4,7 +4,6 @@ package uk.gov.hmcts.reform.em.hrs.controller;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import uk.gov.hmcts.reform.em.hrs.controllers.RootUrlController;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
